feat: Rename and improve BEM application in styled component names
