+++ host-app/src/App.styles.ts
@@ -1,29 +1,27 @@
 import styled from 'styled-components';
 
 export const Container = styled.div`
-  &.host-app__container {
-    min-height: 100vh;
-    color: #fff;
-    font-family: 'Segoe UI', Arial, sans-serif;
-  }
-  & .host-app__main {
-    margin-top: 4rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 4rem;
-  }
+  min-height: 100vh;
+  color: #fff;
+  font-family: 'Segoe UI', Arial, sans-serif;
+`;
+
+export const Main = styled.main`
+  margin-top: 4rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 4rem;
 `;
 
 export const Header = styled.header`
-  &.host-app__header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    gap: 4rem;
-    padding: 2rem 3rem;
-    border-bottom: 1px solid #333;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 4rem;
+  padding: 2rem 3rem;
+  border-bottom: 1px solid #333;
 
     h1 {
       margin: 0;
@@ -36,31 +34,5 @@ export const Header = styled.header`
       border-radius: 4px;
       border: none;
     }
-  }
-`;
 
-export const Selector = styled.section`
-  &.host-app__selector {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-top: 4rem;
-    h2 {
-      margin-bottom: 2rem;
-    }
-    button {
-      margin: 0.5rem 0;
-      padding: 0.75rem 2rem;
-      font-size: 1.1rem;
-      background: #393e46;
-      color: #fff;
-      border: none;
-      border-radius: 6px;
-      cursor: pointer;
-      transition: background 0.2s;
-      &:hover {
-        background: #00adb5;
-      }
-    }
-  }
 `;
